Implement registerUser and according test cases #7067
Currently registerUser fails on the reference server with the message:
"Could not persist the registration: Could not open database: Driver not
loaded Driver not loaded"

This appears to be a problem on the server.
The testcase testRegisterUserNew currently prints the error message
returned by the signal. Once this problem is resolved the line should be
removed.