Fix bug #9682
Introduce new class LoginUser which expands User. LoginUser also takes a
private key and an optional passphrase. Session stores a LoginUser and
uses it to decrypt or sign.

Once a LoginUser has been created the private key cannot be extraced and
may only be used to decrypt or sign.

This change breaks everything which uses login with a non-LoginUser
User.