Changed Appwindow, LoginDialog
Added Entry for starting a chat with a user that isn't in the contact
list. Added an Entry for adding a user to the contact list
In Logindialog: Autohide options when userfield is not empty anymore
made checkbox able to hide, unhide extra options. Autoresize of the
Window when fields are hidden...