Add a couple of flags to our astyle hook
This touches almost every file we have, but I think it's worth it.
Options are now stricter, handling: operator padding (-p), parenthesis
padding (-H, -U), pointer and reference alignment (-W3, -k3), and tabs to
spaces (-c, -s).

A minor problem is that it doesn't handle the throw clause perfectly,
but I can live with that.