Avoid huge QTestLib failure
The QTestLib manual states:
If initTestCase() fails, no testfunction will be executed. If init()
fails, the following testfunction will not be executed, the test will
proceed to the next testfunction.

This is obviously not true for init(). If init() fails, the test case
_is_ executed, and then runs into a QFATAL error because a null pointer
is passed to QSignalSpy. The QFATAL error causes the tests to abort
without running cleanupTestCase or the class destructor, leaving our
server running.

As if that wasn't enough fail, the error messages caused by failures in
init() are misleading because they don't mention init(), only the
corresponding test case.

https://bugreports.qt-project.org/browse/QTBUG-20371