Don't notify chat creator in initChat
This bug was exposed by adding exceptions to the client callback in the
previous commit. The problem here is that LocalChat() is called from
initChat() (server-side), which in turn is called by runInitChat
(client-side). Before runInitChat completes, the chat does not exist as
far as the client is concerned, and therefore calling notifyJoin() makes
no sense.