Avoid duplicate join notification and sync issues
Calling notifyJoin() after adding the participant to the list sent a
duplicate notification to the joining user.

Also, calling notifyJoin() outside of the critical section introduced an
synchronization issue which could result in incorrect users receiving
the notification (for example, if another user joined between the
critical section and the notifyJoin() call).

notifyJoin() was used only in a single spot, so get rid of it entirely.