Prevent inviting a user that is already in chat
This is the second part of #12069.
It prevents both remote and local users from trying to invite a user
that is already present in the chat.

This commit also exposed strange behavior when rethrowing exceptions as
follows:

catch (const Type &e) {
    throw e;
}

In this case, it lead to a complete server lock needing kill -9.

catch (const Type &e) {
    throw;
}

and

catch (const Type &e) {
    throw Type(e.what);
}

both work. Maybe there's something wrong with SDCException's copy
constructor?